IGNITE-25411 Extract tests of --persitence command to separate test
