+++ modules/control-utility/src/test/java/org/apache/ignite/util/GridCommandHandlerTest.java
@@ -24,7 +24,6 @@ import java.io.Serializable;
 import java.lang.reflect.Field;
 import java.nio.file.Files;
 import java.nio.file.OpenOption;
-import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -37,7 +36,6 @@ import java.util.Objects;
 import java.util.Random;
 import java.util.Set;
 import java.util.TreeMap;
-import java.util.TreeSet;
 import java.util.UUID;
 import java.util.concurrent.Callable;
 import java.util.concurrent.CountDownLatch;
@@ -96,14 +94,12 @@ import org.apache.ignite.internal.processors.cache.distributed.dht.GridDhtTxFini
 import org.apache.ignite.internal.processors.cache.distributed.near.GridNearLockResponse;
 import org.apache.ignite.internal.processors.cache.distributed.near.GridNearTxFinishRequest;
 import org.apache.ignite.internal.processors.cache.distributed.near.GridNearTxLocal;
-import org.apache.ignite.internal.processors.cache.persistence.CheckpointState;
 import org.apache.ignite.internal.processors.cache.persistence.GridCacheDatabaseSharedManager;
 import org.apache.ignite.internal.processors.cache.persistence.checkpoint.CheckpointListener;
 import org.apache.ignite.internal.processors.cache.persistence.db.IgniteCacheGroupsWithRestartsTest;
 import org.apache.ignite.internal.processors.cache.persistence.diagnostic.pagelocktracker.dumpprocessors.ToFileDumpProcessor;
 import org.apache.ignite.internal.processors.cache.persistence.file.FileIO;
 import org.apache.ignite.internal.processors.cache.persistence.file.FileIOFactory;
-import org.apache.ignite.internal.processors.cache.persistence.filename.NodeFileTree;
 import org.apache.ignite.internal.processors.cache.persistence.snapshot.DataStreamerUpdatesHandler;
 import org.apache.ignite.internal.processors.cache.persistence.snapshot.IgniteSnapshotManager;
 import org.apache.ignite.internal.processors.cache.persistence.snapshot.SnapshotPartitionsVerifyTaskResult;
@@ -148,7 +144,6 @@ import org.apache.ignite.transactions.Transaction;
 import org.apache.ignite.transactions.TransactionRollbackException;
 import org.apache.ignite.transactions.TransactionTimeoutException;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 import org.junit.Assume;
 import org.junit.Test;
 
@@ -185,7 +180,6 @@ import static org.apache.ignite.internal.processors.job.GridJobProcessor.JOBS_VI
 import static org.apache.ignite.internal.processors.task.GridTaskProcessor.TASKS_VIEW;
 import static org.apache.ignite.testframework.GridTestUtils.assertContains;
 import static org.apache.ignite.testframework.GridTestUtils.assertNotContains;
-import static org.apache.ignite.testframework.GridTestUtils.assertThrows;
 import static org.apache.ignite.testframework.GridTestUtils.runAsync;
 import static org.apache.ignite.testframework.GridTestUtils.waitForCondition;
 import static org.apache.ignite.transactions.TransactionConcurrency.OPTIMISTIC;
@@ -303,361 +297,6 @@ public class GridCommandHandlerTest extends GridCommandHandlerClusterPerMethodAb
 
         if (cliCommandHandler())
             assertContains(log, testOut.toString(), "Command deprecated. Use --set-state instead.");
-    }
-
-    /** */
-    private CacheConfiguration cacheConfiguration(String cacheName) {
-        CacheConfiguration ccfg = new CacheConfiguration(cacheName)
-            .setAtomicityMode(TRANSACTIONAL)
-            .setAffinity(new RendezvousAffinityFunction(false, 32))
-            .setBackups(1);
-
-        return ccfg;
-    }
-
-    /**
-     * Starts cluster of two nodes and prepares situation of corrupted PDS on node2
-     * so it enters maintenance mode on restart.
-     *
-     * @param cachesToStart Configurations of caches that should be started in cluster.
-     * @param cacheToCorrupt Function determining should cache with given name be corrupted or not.
-     */
-    private File startGridAndPutNodeToMaintenance(CacheConfiguration[] cachesToStart,
-                                                  @Nullable Function<String, Boolean> cacheToCorrupt) throws Exception {
-        assert cachesToStart != null && cachesToStart.length > 0;
-
-        IgniteEx ig0 = startGrid(0);
-        IgniteEx ig1 = startGrid(1);
-
-        NodeFileTree ft1 = ig1.context().pdsFolderResolver().fileTree();
-
-        ig0.cluster().baselineAutoAdjustEnabled(false);
-        ig0.cluster().state(ACTIVE);
-
-        IgniteCache dfltCache = ig0.getOrCreateCache(cachesToStart[0]);
-
-        if (cachesToStart.length > 1) {
-            for (int i = 1; i < cachesToStart.length; i++)
-                ig0.getOrCreateCache(cachesToStart[i]);
-        }
-
-        for (int k = 0; k < 1000; k++)
-            dfltCache.put(k, k);
-
-        GridCacheDatabaseSharedManager dbMrg0 = (GridCacheDatabaseSharedManager)ig0.context().cache().context().database();
-        GridCacheDatabaseSharedManager dbMrg1 = (GridCacheDatabaseSharedManager)ig1.context().cache().context().database();
-
-        dbMrg0.forceCheckpoint("cp").futureFor(CheckpointState.FINISHED).get();
-        dbMrg1.forceCheckpoint("cp").futureFor(CheckpointState.FINISHED).get();
-
-        Arrays.stream(cachesToStart)
-            .map(ccfg -> ccfg.getName())
-            .filter(name -> cacheToCorrupt.apply(name))
-            .forEach(name -> ig0.cluster().disableWal(name));
-
-        for (int k = 1000; k < 2000; k++)
-            dfltCache.put(k, k);
-
-        stopGrid(1);
-
-        File[] cpMarkers = ft1.checkpoint().listFiles();
-
-        for (File cpMark : cpMarkers) {
-            if (cpMark.getName().contains("-END"))
-                cpMark.delete();
-        }
-
-        assertThrows(log, () -> startGrid(1), Exception.class, null);
-
-        return ft1.nodeStorage();
-    }
-
-    /**
-     * Test verifies persistence clean command with explicit list of caches to be cleaned.
-     *
-     * @throws Exception If failed.
-     */
-    @Test
-    public void testPersistenceCleanSpecifiedCachesCommand() throws Exception {
-        String cacheName0 = DEFAULT_CACHE_NAME + "0";
-        String cacheName1 = DEFAULT_CACHE_NAME + "1";
-        String cacheName2 = DEFAULT_CACHE_NAME + "2";
-        String cacheName3 = DEFAULT_CACHE_NAME + "3";
-
-        String nonExistingCacheName = DEFAULT_CACHE_NAME + "4";
-
-        File mntcNodeWorkDir = startGridAndPutNodeToMaintenance(
-            new CacheConfiguration[]{
-                cacheConfiguration(cacheName0),
-                cacheConfiguration(cacheName1),
-                cacheConfiguration(cacheName2),
-                cacheConfiguration(cacheName3)
-            },
-            s -> !s.equals(cacheName3));
-
-        IgniteEx ig1 = startGrid(1);
-
-        String port = connectorPort(ig1);
-
-        assertEquals(EXIT_CODE_INVALID_ARGUMENTS, execute(
-            "--host", "localhost", "--port", port,
-            "--persistence", "clean", "caches",
-            nonExistingCacheName));
-
-        assertEquals(EXIT_CODE_OK, execute(
-            "--host", "localhost", "--port", port,
-            "--persistence", "clean", "caches",
-            cacheName0 + "," + cacheName1));
-
-        boolean cleanedEmpty = Arrays.stream(mntcNodeWorkDir.listFiles())
-            .filter(f -> f.getName().contains(cacheName0) || f.getName().contains(cacheName1))
-            .map(f -> f.listFiles().length == 1)
-            .reduce(true, (t, u) -> t && u);
-
-        assertTrue(cleanedEmpty);
-
-        boolean nonCleanedNonEmpty = Arrays.stream(mntcNodeWorkDir.listFiles())
-            .filter(f -> f.getName().contains(cacheName2) || f.getName().contains(cacheName3))
-            .map(f -> f.listFiles().length > 1)
-            .reduce(true, (t, u) -> t && u);
-
-        assertTrue(nonCleanedNonEmpty);
-
-        stopGrid(1);
-
-        ig1 = startGrid(1);
-
-        assertTrue(ig1.context().maintenanceRegistry().isMaintenanceMode());
-
-        assertEquals(EXIT_CODE_OK, execute(
-            "--host", "localhost", "--port", port,
-            "--persistence", "clean", "caches", cacheName2));
-
-        stopGrid(1);
-
-        ig1 = startGrid(1);
-
-        assertFalse(ig1.context().maintenanceRegistry().isMaintenanceMode());
-    }
-
-    /**
-     * Test verifies persistence clean command cleaning only corrupted caches and not touching others.
-     *
-     * @throws Exception If failed.
-     */
-    @Test
-    public void testPersistenceCleanCorruptedCachesCommand() throws Exception {
-        String cacheName0 = DEFAULT_CACHE_NAME + "0";
-        String cacheName1 = DEFAULT_CACHE_NAME + "1";
-        String cacheName2 = DEFAULT_CACHE_NAME + "2";
-        String cacheName3 = DEFAULT_CACHE_NAME + "3";
-
-        File mntcNodeWorkDir = startGridAndPutNodeToMaintenance(
-            new CacheConfiguration[]{
-                cacheConfiguration(cacheName0),
-                cacheConfiguration(cacheName1),
-                cacheConfiguration(cacheName2),
-                cacheConfiguration(cacheName3)
-            },
-            s -> !s.equals(cacheName3));
-
-        IgniteEx ig1 = startGrid(1);
-
-        String port = connectorPort(ig1);
-
-        assertEquals(EXIT_CODE_OK, execute("--host", "localhost", "--port", port, "--persistence"));
-        assertEquals(EXIT_CODE_OK, execute("--host", "localhost", "--port", port, "--persistence", "info"));
-
-        assertEquals(EXIT_CODE_OK, execute("--host", "localhost", "--port", port, "--persistence", "clean", "corrupted"));
-
-        boolean cleanedEmpty = Arrays.stream(mntcNodeWorkDir.listFiles())
-            .filter(f ->
-                f.getName().contains(cacheName0)
-                || f.getName().contains(cacheName1)
-                || f.getName().contains(cacheName2)
-            )
-            .map(f -> f.listFiles().length == 1)
-            .reduce(true, (t, u) -> t && u);
-
-        assertTrue(cleanedEmpty);
-
-        stopGrid(1);
-
-        ig1 = startGrid(1);
-
-        assertFalse(ig1.context().maintenanceRegistry().isMaintenanceMode());
-    }
-
-    /**
-     * Test verifies persistence clean all command that cleans all cache directories.
-     *
-     * @throws Exception
-     */
-    @Test
-    public void testPersistenceCleanAllCachesCommand() throws Exception {
-        String cacheName0 = DEFAULT_CACHE_NAME + "0";
-        String cacheName1 = DEFAULT_CACHE_NAME + "1";
-
-        File mntcNodeWorkDir = startGridAndPutNodeToMaintenance(
-            new CacheConfiguration[]{
-                cacheConfiguration(cacheName0),
-                cacheConfiguration(cacheName1)
-            },
-            s -> s.equals(cacheName0));
-
-        IgniteEx ig1 = startGrid(1);
-
-        assertEquals(EXIT_CODE_OK, execute("--host", "localhost", "--port", connectorPort(ig1),
-            "--persistence", "clean", "all"));
-
-        boolean allEmpty = Arrays.stream(mntcNodeWorkDir.listFiles())
-            .filter(File::isDirectory)
-            .filter(NodeFileTree::cacheDir)
-            .map(f -> f.listFiles().length == 1)
-            .reduce(true, (t, u) -> t && u);
-
-        assertTrue(allEmpty);
-
-        stopGrid(1);
-
-        ig1 = startGrid(1);
-
-        assertFalse(ig1.context().maintenanceRegistry().isMaintenanceMode());
-    }
-
-    /**
-     * Test verifies that persistence backup command to backup all caches backs up all cache directories.
-     *
-     * @throws Exception If failed.
-     */
-    @Test
-    public void testPersistenceBackupAllCachesCommand() throws Exception {
-        String cacheName0 = DEFAULT_CACHE_NAME + "0";
-        String cacheName1 = DEFAULT_CACHE_NAME + "1";
-
-        File mntcNodeWorkDir = startGridAndPutNodeToMaintenance(
-            new CacheConfiguration[]{
-                cacheConfiguration(cacheName0),
-                cacheConfiguration(cacheName1)
-            },
-            s -> s.equals(cacheName0));
-
-        IgniteEx ig1 = startGrid(1);
-
-        assertEquals(EXIT_CODE_OK, execute("--host", "localhost", "--port", connectorPort(ig1),
-            "--persistence", "backup", "all"));
-
-        Set<String> backedUpCacheDirs = Arrays.stream(mntcNodeWorkDir.listFiles())
-            .filter(File::isDirectory)
-            .filter(f -> f.getName().startsWith("backup_"))
-            .map(f -> f.getName().substring("backup_".length()))
-            .collect(Collectors.toCollection(TreeSet::new));
-
-        Set<String> allCacheDirs = Arrays.stream(mntcNodeWorkDir.listFiles())
-            .filter(File::isDirectory)
-            .filter(NodeFileTree::cacheDir)
-            .map(File::getName)
-            .collect(Collectors.toCollection(TreeSet::new));
-
-        assertEqualsCollections(backedUpCacheDirs, allCacheDirs);
-
-        checkCacheAndBackupDirsContent(mntcNodeWorkDir);
-    }
-
-    /**
-     * Test verifies that persistence backup command copies all corrupted caches content to backup directory
-     * but does not touch other directories.
-     *
-     * @throws Exception If failed.
-     */
-    @Test
-    public void testPersistenceBackupCorruptedCachesCommand() throws Exception {
-        String cacheName0 = DEFAULT_CACHE_NAME + "0";
-        String cacheName1 = DEFAULT_CACHE_NAME + "1";
-
-        File mntcNodeWorkDir = startGridAndPutNodeToMaintenance(
-            new CacheConfiguration[]{
-                cacheConfiguration(cacheName0),
-                cacheConfiguration(cacheName1)
-            },
-            s -> s.equals(cacheName0));
-
-        IgniteEx ig1 = startGrid(1);
-
-        assertEquals(EXIT_CODE_OK, execute("--host", "localhost", "--port", connectorPort(ig1),
-            "--persistence", "backup", "corrupted"));
-
-        long backedUpCachesCnt = Arrays.stream(mntcNodeWorkDir.listFiles())
-            .filter(File::isDirectory)
-            .filter(f -> f.getName().startsWith("backup_"))
-            .filter(f -> f.getName().contains(cacheName0))
-            .count();
-
-        assertEquals(1, backedUpCachesCnt);
-
-        checkCacheAndBackupDirsContent(mntcNodeWorkDir);
-    }
-
-    /**
-     * Test verifies that persistence backup command with specified caches copied only content of that caches and
-     * doesn't touch other directories.
-     *
-     * @throws Exception If failed.
-     */
-    @Test
-    public void testPersistenceBackupSpecifiedCachesCommand() throws Exception {
-        String cacheName0 = DEFAULT_CACHE_NAME + "0";
-        String cacheName1 = DEFAULT_CACHE_NAME + "1";
-        String cacheName2 = DEFAULT_CACHE_NAME + "2";
-
-        String nonExistingCacheName = "nonExistingCache";
-
-        File mntcNodeWorkDir = startGridAndPutNodeToMaintenance(
-            new CacheConfiguration[]{
-                cacheConfiguration(cacheName0),
-                cacheConfiguration(cacheName1),
-                cacheConfiguration(cacheName2)
-            },
-            s -> s.equals(cacheName0) || s.equals(cacheName2));
-
-        IgniteEx ig1 = startGrid(1);
-
-        String port = connectorPort(ig1);;
-
-        assertEquals(EXIT_CODE_INVALID_ARGUMENTS, execute("--host", "localhost", "--port", port,
-            "--persistence", "backup", "caches",
-            nonExistingCacheName));
-
-        assertEquals(EXIT_CODE_OK, execute("--host", "localhost", "--port", port,
-            "--persistence", "backup", "caches",
-            cacheName0 + "," + cacheName2));
-
-        long backedUpCachesCnt = Arrays.stream(mntcNodeWorkDir.listFiles())
-            .filter(File::isDirectory)
-            .filter(f -> f.getName().startsWith("backup_"))
-            .count();
-
-        assertEquals(2, backedUpCachesCnt);
-
-        checkCacheAndBackupDirsContent(mntcNodeWorkDir);
-    }
-
-    /** */
-    private void checkCacheAndBackupDirsContent(File mntcNodeWorkDir) {
-        List<File> backupDirs = Arrays.stream(mntcNodeWorkDir.listFiles())
-            .filter(File::isDirectory)
-            .filter(f -> f.getName().startsWith("backup_"))
-            .collect(Collectors.toList());
-
-        Path mntcNodeWorkDirPath = mntcNodeWorkDir.toPath();
-
-        for (File bDir : backupDirs) {
-            File origCacheDir = mntcNodeWorkDirPath.resolve(bDir.getName().substring("backup_".length())).toFile();
-
-            assertTrue(origCacheDir.isDirectory());
-
-            assertEquals(origCacheDir.listFiles().length, bDir.listFiles().length);
-        }
     }
 
     /**
